Short circuiting Callable.__call__ and BoundMethod.__call__
Summary:
Once we flip the switch for BoundMethods, it turns out that looking up the special `__call__` methods on them becomes a major bottleneck.
To mitigate this, I pulled the special case instantiation logic out into a common helper, and immediately returned the attribute when queried.

Note that this does introduce a little duplication in the logic of all of the attribute metadata.  However, I don't imagine the handling of this special attribute will change much down the line, so it seems worth it.

Reviewed By: dkgi

Differential Revision: D20530590

fbshipit-source-id: dbd718d631a2382b2b129fd6d3b58c8bde4fec32