Only look up relevant candidate attributes in instantiate_protocol_attributes
Summary: D19594805 made the ability to look up a single attribute available to the type order.  This diff takes advantage of that to implement a more efficient approach for instantiate_protocol_attributes.  Instead of querying all of the attributes for the protocol and the candidate, we only need to look up all of the attributes for the protocol, and then only look up the matching attributes in the candidate.  This should save time on large candidates being compared against small protocols, as well as paring off some dependency edges.

Reviewed By: dkgi

Differential Revision: D20530340

fbshipit-source-id: 936aa09c21d744593d9a23ac3fe16a0af906a829