Make BoundMethod.__call__ return an anonymous callable
Summary:
Now that we're adding callgraph edges based on the inner callable instead of the one accessed via `BoundMethod.__call__` we can update that function to reflect reality: the modified version of the function really isn't the same function anymore, and thus shouldn't have the same name.  This gets us closer to a long-term aspiration of mine: to maintain a 1-to-1 relationship between function names and their types, which would allow us to store those types indirected through their names.

After this diff, the only remaining modification of named functions is in class-level generic instantiation.

Reviewed By: dkgi

Differential Revision: D20569794

fbshipit-source-id: 9b16bd77a46685551da767c4e1dcfca5cf73ef3f