Do not register type alias for assignment of the form `X = None`
Summary: `X = None` should not be treated as a type alias assignment -- `None` is a value, not a type. The corresponding type for `None` is `NoneType`, but there's no way to explicitly write out `NoneType` in Python (they probably want to prevent people from subclassing it, etc.).  The fact that one could write `None` as param/return type is just a compromise to the `NoneType` situation.

Reviewed By: arthaud

Differential Revision: D20565327

fbshipit-source-id: c752ea9986248365b3f35226747cd386b2296fa8