Only returning return type from signature selection
Summary:
With D20614211, we are no longer using anything from signature selection but the instantiated return type of the selected overload.  This means we can stop emitting the parameters we no longer use, making the interface a lot cleaner, and leaving the door open for not instantiating the parameters at all.

Most of the work of this diff was for retrofitting annotatedSignatureTest to check a single return type rather than a whole overload.

Reviewed By: sinancepel

Differential Revision: D20614155

fbshipit-source-id: 48fec0e611be50e1b68ffb53946b1cb087305e17