Small improvement on watchman logging, part II
Summary:
Let's avoid spamming the rage log with hundreds of filenames when one does a rebase in fbsource.

What I'm trying to achieve here is that If the number of changed files is greater than a threshold (I propose 30 but open to other suggestions), we simply stop printing the additional items.

Reviewed By: arthaud

Differential Revision: D20661581

fbshipit-source-id: 9973954ee0151d89f763e9fd6d202863e6d08489