Handle malformed ECONNRESET exceptions
Summary: In theory the exception that's supposed to be emitted because of a RST packet is supposed to be `Unix.Unix_error (Unix.ECONNRESET, _, _)`. However for reasons unknown, sometimes the OCaml networking stack emits a `Sys_error "Connection reset by peer"` instead. The hacky fix herein is just to match for both of these.

Reviewed By: dkgi

Differential Revision: D20661780

fbshipit-source-id: 541da675e33059d3f9804d5a07b1d594a2ff0c30