Fix "bad file descriptor" concurrency bug
Summary:
D20323838 did most of the work of fixing the race condition between the computation and request handling threads, vis a vis closing a file descriptor that we wanted to `select` on.
Unfortunately there was a small mistake that limited its effectiveness: we were doing the closing of the sockets **after** we fetched the list of sockets we were going to `select` on.
This meant the following scenario was still possible

```
request handling thread                   computation thread
Mutex.critical_section
  lock
  ~f:(fun () -> !raw_connections)
# json_sockets = [42]

                                          remove_json_socket(42)
                                          # marks 42 as a "socket_to_close"
Connections.close_json_sockets
   ~connections
# Does a Socket.close(42)
Unix.select(42)
# Error!
```

Luckily the fix here is simple, just do the closing before fetching the list to `select` on.  This means that any `removes` that happen after the `close` will only be actually closed on the next cycle.

Reviewed By: MaggieMoss

Differential Revision: D20676050

fbshipit-source-id: 992ff83afb71f21f44337a4f8837d1f07f6be0cf