Supporting functions that pass a subset of their arguments on to another function with ParameterSpecification
Summary:
The last feature described in PEP 612 that we have yet to implement is the following example:
```
def call_n_times(
    __f: Callable[TParams, None],
    __n: int,
    *args: TParams.args,
    **kwargs: TParams.kwargs,
) -> None:
    for x in range(__n);
        __f(*args, **kwargs)
```
The idea is that the type call_n_times has two positional-only arguments, and then the remaining arguments are determined by the type of the callable passed in as the first argument.

Effectively this is implementing the call semantics for `Callable<TParams>[Concatenate[contains_tparams, TParams], something]`.

The idea of this implementation is to split one signature selection into two:
* a first one of the head of the arguments against the head of the parameters
* then extract a parameter set from the constraints gathered in that first selection
* then a second selection of the remaining arguments against those "generated" arguments

One especially cool thing about this approach is that by mapping the constraints set from the first selection through an `extract_partial_solution` we get handling of overloaded functions more or less for free :).

The other reason I was interested in closing this final hole is that integrated with my earlier BoundMethod work, this instantly gets us the solution to the PyTorch Module problem.
The core issue of typing the PyTorch Module class is that the signature of a method on an abstract superclass is being determined by the signature of different method on children of that class.  The best solution I came up with earlier was making the abstract superclass generic with regards to the signature of the child method, but that required respelling the type of the child method into the generic parameters of the parent, which was doubly complicated by the intentional lack of a "literal" syntax for parameter specifications.

Instead, by making self argument just another argument (D20495185) and implementing the above functionality, we can extract the child's signature via a generic protocol, then "upherit" those parameters for ourselves :).
See typeVariableTest.ml for a working example of this.

Reviewed By: dkgi

Differential Revision: D20659961

fbshipit-source-id: 971982e95e898b1b621498b9f8a4105b65bec4ce