Getting rid of most usage of Pervasives.compare/eq
Summary:
To prepare for newer version of ocaml, and for speed, we should use as less
polymorphic equality as possible. This diff removes almost all of the
Pervasives.compare/eq usages.

A couple are left, mostly involving GADT or very big type structure.
Feel free to try and kill the remaining ones. You can find them by searching for `Stdlib.compare` or `Stdlib.(=)`

Differential Revision: D20691916

fbshipit-source-id: 975f6f14e709e83191ad4fb6c0896efe51fe6df3