Stop using join in solve_less_or_equal
Summary:
I'm trying to pull solve_less_or_equal out of TypeOrder, and it would be much easier to do this if solve_less_or_equal did not depend on anything we'll be leaving behind in TypeOrder.
There is currently only one use of `join`, and it can be easily replaced with a `Type.union` in this case.

For almost all types `Type.union [X, Y]` = `TypeOrder.join X Y` and for the other ones, `Type.union [X, Y]` < `TypeOrder.join X Y` (aside: for this reason I think it would be possible to kill join and just replace it with Type.union, I've tried before and I'll likely try again).

This makes this almost always a non-functional change.

Reviewed By: grievejia

Differential Revision: D20716010

fbshipit-source-id: 39f6b67e04d64058b7081fdbf96992ea20d82757