Separating solve_less_or_equal and its cronies out into its own module
Summary:
After discussing pain points in the backend with dkgi, I decided that one way to make this stuff easier to understand was to provide a better api to deal with our representation of `TypeConstraints.t list`.  This is supposed to represent a disjunction of multiple constraints objects, but interacting with it is definitely a bit confusing.  The core idea of this refactor is to hide that complexity behind a TypeConstraintsSet module.

The division of labor here is as follows:
TypeConstraints: data type is a conjunction of different bounds on particular variables.  Interface is adding a series of upper and lower bounds to **variables**, and then solving those bounds out to get values for type variables.
TypeConstraintsSet: data type is a disjunction of the TypeConstraints data type.  Interface is a sequence of solve_less_or_equals, AKA a system of inequalities on **types**, and then solving that system.
TypeOrder: ends up being pretty small.  just basically defines `join` and `meet`.  More importantly for our purposes here it ties the mutual module recursion knot by passing itself to construct TypeConstraints, which it then passes onto TypeConstraintsSet, which it then is itself defined on.

Questions:
* Naming.  Not completely happy with TypeConstraintsSet.  Other ideas: rename TypeConstraints to VariableConstraints or VariableBounds, and then name TypeConstraintsSet TypeConstraints?
* More generally: does having the main interface you deal with hide the `TypeConstraints.t list` seem like it would make things more comprehensible?

Reviewed By: dkgi

Differential Revision: D20716753

fbshipit-source-id: cd9f902138ecd367be32d529fd609e3e449f915d