support custom `__getitem__` implementations for unpacking tuples
Summary:
Basically, what happens in `x, y, z = t` is: `t.__getitem__(index)` gets called with increasing indices until a IndexError is reached, and that list is mapped to `x, y, z`.

We special-case this for tuples, but don't handle the general case.

Let's add support for custom classes as well in this diff. We don't handle arity, just as we don't throw an exception for `a = []; a[0]`.

Reviewed By: dkgi

Differential Revision: D20772317

fbshipit-source-id: 856f57dc28e1724ca858204da9cc6c8bb3354cac