add support for adjusting positions of parsed model parameters
Summary:
Context: Pysa supports omitting optional parameters from your models. This is so that for things like:

```
def foo(x, optional_1=..., ..., optional_50=...)`
```

you don't get churn when someone adds `optional_51 = ...` to the end of the define.

The problem: What if the model writer omits  `optional_25`? In that case, the positions for any parameter after that one would be off. To fix that issue, get the actual positions of named parameters from the source code and apply those.

The reason we used an integration test (that demonstrates the attach and sink models correctly pointing to the same parameter/position now) instead of the model test is because we ignore parameter positions in all of Pysa testing, and changing that infra seemed heavy handed for just this.

Reviewed By: dkgi

Differential Revision: D20779423

fbshipit-source-id: 3641e7af0131d2625b827ac07e8ef3f005e08d0a