Fixing handling of variable and keywords arguments in simulate signature select
Summary: *args and **kwargs arguments are entirely optional, so when determining whether a callable is a behavioral subtype as another, we should feel free to discard the prospective subtypes variable and keywords parameters when we are done using them.  As of now, we only do this when we are completely out of pseudo-arguments, preventing us from handling a scenario where the supertype callable has a kwargs but no args, and the subtype has both.

Reviewed By: dkgi

Differential Revision: D20779127

fbshipit-source-id: 963a388062db827ce9f762caf037ac06c073aefd