ConstraintsSet.solve_less_or_equal -> ConstraintsSet.add
Summary:
One of the aspects of simplifying the interface of ConstraintsSet is renaming its core action to something more evocative.  I think that a reasonable option here is to just call it add, since we're just adding a constraint to a set of constraints.

This stack gradually pushes this renaming out through the layers of clients that use this interface.

Reviewed By: dkgi

Differential Revision: D20780646

fbshipit-source-id: b7d8b223cc43f303ed0f7a77402171978c684f53