Exposing an interface to solve `ConstraintsSet.t`s
Summary:
The other set of usages of the underlying list of disjunctions of type constraints is the process of solving.

In general the pattern is
```
List.filter_map a_list ~f:(TypeConstraints.solve ~order)
|> hd
>>| use_solution
```

This diff offers an abstraction of that pattern of the form
```
ConstraintsSet.solve constraints_set ~order
>>| use solution
```

I'll gradually replace existing instances of the first pattern with the second in the next few diffs.

Reviewed By: dkgi

Differential Revision: D20799943

fbshipit-source-id: 044a1bab592f45c451462a9e66827044acf53846