Basing lambda inference logic off of new interface
Summary:
The third out of the four usages of extract_partial_solution we have to deal with comes from our special case handling of lambda expressions.  The core idea here is that we need to get the solved type for the parameter of the lambda, then use that to infer the type of the lambda body, then use that as a constraint itself.

This usage doesn't care about multiple possible solutions, so it turns out we can adapt it slightly to use the simpler interface.  The idea here is that once we've gotten a solution for that variable, we can use it to make other constraints, but we then need to "commit" to that choice, to ensure that later addition of bounds won't slide it to another point in the range in a way that was incompatible with the choices we made.

This "committal" could have been done (in this case) with two `LessOrEqual`s, but I thought it was more evocative (and was helpful for the next example) to label a special `VariableIsExactly` constraint here.

Reviewed By: sinancepel

Differential Revision: D20798298

fbshipit-source-id: b23c6b7ee0e59d14b3ff3cb3e555dc8e8ebc7a85