Exposing and using a new interface for the last usage of extract_partial_solution
Summary:
The last, and most complex usage of extract_partial_solution is the one introduced by D20659961 to handle signatures of the following form:
```
def call_this_function(f: Callable[TParams, TReturn], /, *args: TParams.args, **kwargs: TParams.kwargs) -> TReturn: ...
```
The wrinkle here is that we'd like to handle overloads here, i.e. get different return types based on the arguments in the tail, corresponding to the overloads of `f`.

This means we can't do the "single solution" simplification that we used for D20798298, and thus need to, for the first time, expose the fact that we can have more than one solution for the constraints set.  We could certainly expose that fact in general, but I think in the simple cases we don't need to worry about it, and its better to just hide that fact.  Accordingly, I only open up the door a crack, just for this use case to slip in.

Also note that we have to use the same "commital" strategy as D20659961, which in effect selects the remainder constraints we had before.

Reviewed By: dkgi

Differential Revision: D20801777

fbshipit-source-id: 6bf09a0337c6cdee905674c9bc7626ebb5b4b941