refactor special call handling in forward analysis to its own function
Summary: Problem: The special cases for calls were split between the analyze_call function and the pattern matching in the high-level anaylze_expression code. This meant that it wasn't clear where a piece of logic would be. To mitigate this, move everything to analyze_call, and hide the previous body of analyze_call inside to ensure it's not reachable by external callers.

Reviewed By: fahndrich

Differential Revision: D20841064

fbshipit-source-id: cb4884f831e2f76b3a5f1a0a24d10b0b624c7652