Extract dataclass for arguments to `servers`.
Summary: Problem: When one command calls another, such as Incremental calling Start, it passes on its `arguments` as `argparse.Namespace`. This means that we may forget to insert an argument needed by the other command, such as `--no-watchman` for Start. This is not caught by the typechecker and fails at runtime.

Reviewed By: grievejia

Differential Revision: D20875068

fbshipit-source-id: 7a86f8d95441cc5fd95ce472aad9e2a09da7c9b1