Replace UserSpecified with CustomUserControlled
Summary:
I noticed that most of the codebase uses `UserControlled` but some documentation and integration tests use `UserSpecified`.

This change updates `UserSpecified` to be `CustomUserControlled` to be explicit that we want to differentiate from `UserControlled` within the Pysa tutorials. In other parts of the documentation and integrations tests `UserSpecified` is replaced with `UserControlled`.

Reviewed By: gbleaney

Differential Revision: D20856273

fbshipit-source-id: 4178b091e3d6febb2a0452f3cf61987179e849ff