Remove `--buck-builder-debug` option
Summary:
IMO it is not necessary to have a separated debug flag just for the buck builder. If one wants to debug buck builder, he should just do so directly with the buck builder target itself instead of doing it indirectly via the pyre client.

`tbgs` suggests that the flag is also not used anywhere on sandcastle. Let's just merge this option into the pyre debug option.

Reviewed By: pradeep90

Differential Revision: D20935100

fbshipit-source-id: 0603823f8ac5daf93df299028e010a1eb7ecf131