Do not spawn and overwrite logging handler
Summary:
This fixes the cleanup issues we were experiencing in the logger, where for some reason non-sticky output was not getting erased if it was the last thing printed (and we were often appending the next cli prompt to the tail of the previous pyre info log).

Cleanup wasn't working because it somehow believed `_active_lines` attribute was always 0, no matter whether the previous emit clearly set it to `1`.

Turns out this is because we are initializing a logger in `facebook/client/pyre.py`, and another logger in `client/pyre.py`. The second initialize overwrote the global `__handler` object in log.py, meaning logger object we were emitting messages on was *not the same* object as the one we were calling cleanup on! :'(

killme

Reviewed By: grievejia

Differential Revision: D20969605

fbshipit-source-id: d5ab7360b1b0fe20843dea97fe231c265c357ea6