start parsing combined source rules
Summary:
In order to be able to express correct rules, we also needed to introduce the idea of a "triggered partial sink" - we don't want to create false flows here.

How this ties in to the high-level analysis: When doing forward analysis, we'll start reading flows to partial sinks, and create partial sinks for the other piece of the pair as appropriate.

Reviewed By: fahndrich

Differential Revision: D20971695

fbshipit-source-id: 750742c2aa93c541f5b288e55aebe1d9ccc91b10