introduce the `multi sink` kind and validate during parsing of rules
Summary:
Context: Sinks that have associated labels with them are special, and we should force the specifier
of these rules to deal with the complexity - a sink can either be a multisink, in which case the only valid
use of it is within a partial sink, or it can be a regular sink, in which case it can be used for regular rules.
Let's start enforcing this in code.

Note that I chose to introduce multi sinks as a purely parsing artifact instead of something we keep around in the code, and refuse to
parse rules that don't have the right kind of sink. This made pragmatic sense to me - this means that
the surface of multi sinks is limited to user-facing syntax, without us having to handle more cases in the analysis.

Reviewed By: fahndrich

Differential Revision: D20997864

fbshipit-source-id: 6d4c1ed1bd8318ad4b1febff9d4b4fb0bc1da33f