Directly support arguments without expressions
Summary: After cleaning up some unnecessary uses of expressions in the stack below, we can now directly support arguments (inside of signature_select at least) that have no expression.  We already were doing this for self_argument, but instead of actively handling that, we instead just used an ellipsis with a `Location.any`.  To elegantly support "synthetic" calls as described in D21025126 we need to support arguments without expressions.

Reviewed By: sinancepel

Differential Revision: D21023290

fbshipit-source-id: 0f4cd726575363f71ea3327b9f371a2e6801f11a