Use resolved signature select to avoid re-resolving arguments
Summary:
The first of the two major reasons I embarked on this refactor is to support this perf optimization:
Previously we were resolving every argument to every function (at least) twice.  This was because we had to resolve it once in TypeCheck to get the errors that might have been from those expressions, and once in the signature select to get the type to compare against the argument type.
This was rendered unnecessary by the ability to pass the already resolved type into signature select, avoiding that second resolve.

Reviewed By: sinancepel

Differential Revision: D21026994

fbshipit-source-id: 842917e4baf1bb21fb1165a1e78e7e4db7f8a65c