Inlining GlobalResolution.class_definitions into ModelParser
Summary:
Unmanaged caches are a potential source of inconsistencies.  This one is currently not causing any, since its currently only used in taint analysis, which is not incremental.  However, locating it inside of globalResolution is tempting people to use it with an implicit guarantee that it, like other things in globalResolution, is dependency tracked.

Instead, it seems reasonable to locate it directly where it's used, where its clear that such caches are safe.

Reviewed By: grievejia

Differential Revision: D21053357

fbshipit-source-id: 27fd115cfc4990ea813dabfe0f1059720385bd40