start triggering partial sinks
Summary:
This diff was the point of the stack of changes. This diff handles the generation of triggered sinks for a parameter when the other one in the multi-sink issue has a flow.

We've cut a bunch of corners at the moment to keep the problem tractable:

Some problems off the top of my head:

* We don't handle the case where both flows exist at the same call.
* We don't store traces for both flows properly at the moment - you'll only see the source flow to the triggered sink.
* We don't propagate partial sinks when both parts flow to the parameters in backwards analysis.

I'll have diffs for all these problems stacked on top.

Reviewed By: fahndrich

Differential Revision: D21034685

fbshipit-source-id: ee5fdf19825415c9140dae2f4b257a0b89259f06