modify order of parsing configurations to merge all sources & sinks before rules
Summary:
The goal here is simple: If users split up configs into multiple files, we'd like them to not have
to redefine sources & sinks in each file. By changing the order in which we parse sources to eagerly merge fields as we parse them,
we're able to make sure users don't need to write boilerplate.

Reviewed By: fahndrich

Differential Revision: D21056902

fbshipit-source-id: 608d2ea6a7d0827a744382c1ea3b50fb6383a19b