Making class decoration produce normal methods
Summary: Class decoration was implemented to produce either bare callables or BoundMethods depending on whether it was being accessed through the class or through the instance. However, this handling already exists inside of instantiate_attribute, as long as this emitted `Method`s instead of `Attributes`, which has the advantage of being correct.

Reviewed By: dkgi

Differential Revision: D21073313

fbshipit-source-id: b85b29f3efdfb5cf40b50d6172cc1771209934e0