Mark class properties as initialized OnClass, not as ClassVars
Summary:
The next part of my quest to clean up the semantics of class_attribute is class properties.  Again, this actually has nothing to do with `ClassVar` and in general we don't have a model for the instance writability of these properties.  That model will fall out of a complete implementation of descriptors.

The correct thing to infer from a class property is that it's accessible from the class, unlike other properties which are only accessible on the instance (in the form that we surface them now).  This will be the significance of OnClass/OnlyOnInstance, so thats actually how we should be surfacing that difference.

Reviewed By: dkgi

Differential Revision: D21058002

fbshipit-source-id: 3e19bd22dc86e6fdcbaf9fca748407528491b39d