Marking enums as ReadOnlys, not ClassVars
Summary: I'm trying to clean up the use of the `class_attribute` property to mean one thing: it was declared as a `ClassVar`.  Currently it means a bunch of conflicting things.  One of these things is "is an enum property".  I believe this was trying to make it so you'd know you could access it from the class, but that's not how that property is being used.  Instead we should require that it be read-only (from both the class and the instance), which it actually is.

Reviewed By: grievejia

Differential Revision: D21057436

fbshipit-source-id: 9da8c30507effea362d73902a0245d5a9a1d2802