Enforcing compatible typing on (valid) assignments to Finals
Summary:
We previously were completely skipping assignment type validation when that attribute was final.  This was based off the presumption that these assignments would either be of the form `x: Final = literal` and thus couldn't be incorrect, or were external assignments that were already invalid for another reason.  However, this meant we blindly accepted things like
```
class Foo:
  x: Final[int] = "A"
```
or
```
dataclass(frozen=True)
class D:
  x: int = "A"
```

This diff preserves some of the original behavior in that invalid external assignments will continue just to have the `Final` error, and not a type error as well (This is particularly important because Final implies we type the variable as `Literal`, which would otherwise lead to some strange errors).
However, when the finality isn't being violated, we now correctly also run the type check.

Reviewed By: pradeep90

Differential Revision: D21057409

fbshipit-source-id: 4bf5bead73ae19ae1252156ff210b197bb9ba0de