Remove confusing ClassVar error
Summary:
The semantics of `ClassVar` are defined in pep 526 to be
> Here ClassVar is a special class defined by the typing module that indicates to the static type checker that this variable should not be set on instances.

We had a secondary message on missing attribute errors that suggested that it instead was *required* to have a variable be accessible from the class.  This is neither true in our implementation, or in the spec.  Instead, we're lacking enforcement of accessing instance attributes through the class, but those error messages will be separate from this.

Reviewed By: dkgi

Differential Revision: D21058395

fbshipit-source-id: 2afb2037e99baaa072d1c7117d3020052409b379