Fix client hanging when malformed local root is used
Summary:
Running the following command would throw `pyre` into an infinite loop:

```
> pyre -l //foo/bar check
```

Root cause of this problem is how `find_root` is implemented: the loop would only terminate if `current_directory` becomes `/`. But for paths that start with `//`, `os.path.dirname` will eventually transform `current_directory` into `//`. And since `os.path.dirname('//') = '//'`, we are never going to get `/`  and thefore will just busy loop there forever :(

The fix is to change the loop termination condition -- we should break the loop not on `/`, but on having `os.path.dirname` returning the same path as `current_directory`.

Reviewed By: pradeep90

Differential Revision: D21082571

fbshipit-source-id: 43937cf669aeb0aaaf9b08ca0e013a4114932eb8