Correctly flatten Union[Optional[Union[x, y]], x] to Optional[Union[x, y]]
Summary:
Union flattening is critical for performance since otherwise we can get tons of layers of unions of unions of unions, which gradually grinds everything to a halt.  We already were attempting to handle Optionals as another form of union, but this handling didn't support when one of the arguments to `Type.union` was already a `Optional[Union[...]]`.

This diff adds this handling, which allows things that never terminated in a reasonable amount of time to terminate.

Reviewed By: sinancepel

Differential Revision: D21079386

fbshipit-source-id: eddbe1d955239842f64ae07a1ac2a3c05c8f3576