Ensure logging cleanup happens when argparse exits
Summary:
Looks like it's hard to have any control over the way argparse exits/errors either when a bad command is passed in or when --help is called.

Adding a try/except block around the parsing does not help, but we can override the exit and error handling of the argument parser to ensure logging gets cleaned up and we return the cursor properly to the user.

Reviewed By: dkgi

Differential Revision: D21044303

fbshipit-source-id: abad990c88cf187ecb87e06476cb629db949e42e