Make result error message available for logging in pyre client
Summary:
RFC: This would help us track results that contain error messages; currently only set up for query which was specifically requested.

Other options considered included trying to special-case pyre query to pass back a response rather than the current model of logging to standard out and dropping it, but this seemed messier; also thought about creating a separate scuba table for pyre query that could better track its error cases. A lot of the info there would need to be duplicated though, and it seems more generically useful to revive the existing but currently empty error_message column of the pyre usage table.

Not sure if this is the best design, open to other ideas.

Reviewed By: dkgi

Differential Revision: D21104521

fbshipit-source-id: 26239ef72d0723d70d4f18827533c76e44a1ff41