Assert negation of test on `while` `else` block
Summary:
Now that we're correctly always routing through the `else` block (D21197188), we can take advantage of that block to accomplish something: putting in the negated assumption.  We know we only hit the `else` block when the test fails, so we can assert its negation just as we do for normal conditionals.  This elegantly gives us a lot of desirable properties downstream:
* going through the `else` pathway on `while True` loops are correctly inferred to be impossible (i.e. State.Bottom), so assignments etc. preceding a `break` are honored after the loop.  This also means any type errors following an unbroken `while(True)` are correctly suppressed.
* if you actually have a `else` branch, we can refine types to be helpful inside of that branch.

Reviewed By: grievejia

Differential Revision: D21196690

fbshipit-source-id: 4c97be6541bdcfca7a55233a11efa68e373eae3c