Remove spurious edge from test directly to end of while loop
Summary:
The semantics of a while loop in python are the following:

1: <enter> run test, if pass go to 2, if fail go to 3
2: run main block, on continue go to 1.  on break go to 4, otherwise go to 1 on completion.
3: run `else` block, go to 4
4: <exit>

There is no situation where we go directly from 1 to 4 without passing through 3.  Normally this mistake doesn't matter because people almost never use `else` blocks on `while`s (so 3 disappears and 1 in fact does go straight to 4), but it will become significant once we start synthesizing an assert into the `else` block in D21196690.

Reviewed By: grievejia

Differential Revision: D21197188

fbshipit-source-id: dbf00acaff4541e5d982501943cef8a7acc141c6