Restore `buck_mode` flag for buck builder
Summary:
We used to thread through a `buck_mode` argument from the client to the buck builder in D19857066. But the semantics was lost in a refactoring diff D21095683. This causes pyre to break in `ovrsource` as the repo relies on a specific buck mode. Let's fix it.

I've also made `buck_mode` a non-optional argument so that this kind of issue can be easier to discover in the future.

Reviewed By: pradeep90

Differential Revision: D21215035

fbshipit-source-id: 0da7cf8687fd7507c469fef20216780e9ef3ec8e