Stop using DummyContext in inference
Summary: I encountered some weird inconsistencies between infer and check behavior when I tried to use the enclosing define to clean up some errors.  This was due to the implicit use of the dummy context.  This is unnecessary for inference, because we do in fact maintain a real context, which we can pass in.

Reviewed By: dkgi

Differential Revision: D21193929

fbshipit-source-id: 8ae627721fdb9c101f79413738d13532dad19222