Make use of dummy context explicit and optional
Summary:
`TypeCheck.resolution` and `TypeCheck.resolution_with_key` are currently less authoritative than they appear.  Because all it has access to is a global resolution, it needs to make up its own dummy context, which locates the type checking inside a nonexistent empty toplevel function.  Because some elements of type checking depend on the nature of that containing function, this is not an insignificant choice.  However, because its hidden away inside the implementation, callers aren't aware of the choice that they're making.

This diff restructures things such that clients now need to explicitly pass in a context, making visible the fact that they're using a dummy context.  This also opens the door for some clients to stop doing this (as I do in my follow ups).

Reviewed By: dkgi

Differential Revision: D21193752

fbshipit-source-id: 3dc8e4c48259d057590c6b76795d7488eda5d66a