Use `typing.StaticMethod` to remove special casing for `@staticmethod`
Summary:
Now that we have a class definition with the real descriptor behavior encoded on it, we can now exploit it to treat staticmethods as attributes, not special cased as Methods.  With all the descriptor infra, this is as simple as having create_attribute emit a typing.StaticMethod[callable_type] attribute instead of a callable_type method.
We then exploit the saved uninstantiated attribute to continue to support the `static` interface on `AnnotatedAttribute`.

Reviewed By: grievejia

Differential Revision: D21190693

fbshipit-source-id: 43fe84097b70a690b30a6ec43c625fcde07de455