Adding fake `typing.ClassMethod` class
Summary:
Unfortunately for our plans to remove the special casing around `classmethod`, the existing class definition for `classmethod` is fundamentally insufficient:
```
class classmethod(object):  # Special, only valid as a decorator.
    __func__: Callable[..., Any]
    if sys.version_info >= (3,):
        __isabstractmethod__: bool

    def __init__(self, f: Callable[..., Any]) -> None: ...
    def __new__(cls: Type[_T], *args: Any, **kwargs: Any) -> _T: ...
    def __get__(self, obj: _T, type: Optional[Type[_T]] = ...) -> Callable[..., Any]: ...
````

Without being a Generic class, it can't store the type of the contained function as anything other than `Callable[..., Any]`.
We could PR typeshed to fake it as Generic, but that has problems of it own, as we've discovered with other classes which are generic in typeshed but not in the runtime.

The solution for similar dilemmas before is to make a capitalized variant of the type in `typing`, where we can make it whatever we want.  This diff does that for `ClassMethod`.

Reviewed By: grievejia

Differential Revision: D21186533

fbshipit-source-id: 16938963cb5318d0bf1069ab119895a379d25790