Augmenting AnnotatedAttribute with the uninstantiated type
Summary:
We'd like to kill the `static` property of AnnotatedAttribute.t when moving its implementation over to the `StaticMethod` descriptor.  However, `static` was actually being used once by `TypeCheck`, so we need to be able to distinguish between `static` and non-static methods even when moving to the new model.
There are a variety of solutions to this, but I think the most valuable long term is to base it off of the uninstantiated type, which needs to be stored as well.  The idea here is that override checking etc. should really operate on the uninstantated type, as that correctly captures all of the behaviors that a class offers.

NOTE: This needs to temporarily be an optional type until we rectify the situation with `property`.

Reviewed By: grievejia

Differential Revision: D21190577

fbshipit-source-id: b7447b791c032b3e3dde73277524ab232623a6b5